requirements.txt
fastapi
uvicorn
gunicorn
jinja2
python-multipart
azure-core
azure-cosmos
python-dotenv